(PC-11652) Move ubble subscription functions in separate folder
