+++ api/src/pcapi/core/subscription/api.py
@@ -2,10 +2,7 @@ import datetime
 import logging
 import typing
 
-import flask
-
 from pcapi import settings
-from pcapi.connectors.beneficiaries import ubble
 from pcapi.core.fraud import exceptions as fraud_exceptions
 import pcapi.core.fraud.api as fraud_api
 import pcapi.core.fraud.models as fraud_models
@@ -227,95 +224,6 @@ def _send_beneficiary_activation_email(user: users_models.User, has_activated_ac
         )
     else:
         accepted_as_beneficiary_email.send_accepted_as_beneficiary_email(user=user)
-
-
-def start_ubble_workflow(user: users_models.User, redirect_url: str) -> str:
-    content = ubble.start_identification(
-        user_id=user.id,
-        phone_number=user.phoneNumber,
-        first_name=user.firstName,
-        last_name=user.lastName,
-        webhook_url=flask.url_for("Public API.ubble_webhook_update_application_status", _external=True),
-        redirect_url=redirect_url,
-    )
-    fraud_api.ubble.start_ubble_fraud_check(user, content)
-    return content.identification_url
-
-
-def is_ubble_workflow_restartable(fraud_check: fraud_models.BeneficiaryFraudCheck) -> bool:
-    if fraud_check.type != fraud_models.FraudCheckType.UBBLE:
-        return False
-
-    ubble_content: fraud_models.ubble_models.UbbleContent = fraud_check.source_data()
-    if ubble_content.status == fraud_models.ubble_models.UbbleIdentificationStatus.INITIATED:
-        return True
-    return False
-
-
-def update_ubble_workflow(
-    fraud_check: fraud_models.BeneficiaryFraudCheck, status: fraud_models.ubble.UbbleIdentificationStatus
-) -> None:
-    content = ubble.get_content(fraud_check.thirdPartyId)
-
-    if not settings.IS_PROD and fraud_api.ubble.does_match_ubble_test_email(fraud_check.user.email):
-        content.birth_date = fraud_check.user.dateOfBirth.date() if fraud_check.user.dateOfBirth else None
-
-    fraud_check.resultContent = content
-    pcapi_repository.repository.save(fraud_check)
-
-    user = fraud_check.user
-
-    if status == fraud_models.ubble.UbbleIdentificationStatus.PROCESSING:
-        user.hasCompletedIdCheck = True
-        pcapi_repository.repository.save(user)
-        subscription_messages.on_review_pending(user)
-
-    elif status == fraud_models.ubble.UbbleIdentificationStatus.PROCESSED:
-        try:
-            fraud_api.ubble.on_ubble_result(fraud_check)
-
-        except fraud_exceptions.BeneficiaryFraudResultCannotBeDowngraded:
-            logger.warning(
-                "Trying to downgrade a beneficiary that already has been considered OK", extra={"user_id": user.id}
-            )
-
-        except Exception:  # pylint: disable=broad-except
-            logger.exception("Error on Ubble fraud check result: %s", extra={"user_id": user.id})
-
-        else:
-            if fraud_check.status != fraud_models.FraudCheckStatus.OK:
-                handle_validation_errors(
-                    user=user,
-                    reason_codes=fraud_check.reasonCodes,
-                )
-                subscription_messages.on_ubble_journey_cannot_continue(user)
-                return
-
-            try:
-                on_successful_application(
-                    user=user,
-                    source=BeneficiaryImportSources.ubble,
-                    source_data=fraud_check.source_data(),
-                    eligibility_type=fraud_api.get_eligibility_type(fraud_check.source_data()),
-                    third_party_id=fraud_check.thirdPartyId,
-                    source_id=None,
-                )
-
-            except Exception as err:  # pylint: disable=broad-except
-                logger.warning(
-                    "Could not save application %s, because of error: %s",
-                    fraud_check.thirdPartyId,
-                    err,
-                )
-            else:
-                logger.info(
-                    "Successfully created user for application %s",
-                    fraud_check.thirdPartyId,
-                )
-
-    elif status == fraud_models.ubble.UbbleIdentificationStatus.ABORTED:
-        fraud_check.status = fraud_models.FraudCheckStatus.CANCELED
-        pcapi_repository.repository.save(fraud_check)
 
 
 def has_completed_profile(user: users_models.User) -> bool:
@@ -548,6 +456,7 @@ def handle_validation_errors(
             subscription_messages.on_duplicate_user(user)
 
 
+# TODO (viconnex): move in a dms folder
 def start_workflow(
     user: users_models.User, thirdparty_id: str, content: fraud_models.DMSContent
 ) -> fraud_models.BeneficiaryFraudCheck:
