+++ api/tests/core/subscription/test_api.py
@@ -1,5 +1,4 @@
 from datetime import datetime
-import json
 from unittest.mock import MagicMock
 from unittest.mock import patch
 
@@ -24,10 +23,6 @@ from pcapi.models.beneficiary_import import BeneficiaryImport
 from pcapi.models.beneficiary_import import BeneficiaryImportSources
 from pcapi.models.beneficiary_import_status import ImportStatus
 import pcapi.notifications.push.testing as push_testing
-
-from tests.core.subscription.test_factories import IdentificationState
-from tests.core.subscription.test_factories import UbbleIdentificationResponseFactory
-from tests.test_utils import json_default
 
 
 @pytest.mark.usefixtures("db_session")
@@ -288,108 +283,6 @@ class EduconnectFlowTest:
         assert response.status_code == 204
         assert user.roles == [users_models.UserRole.UNDERAGE_BENEFICIARY]
         assert user.deposit.amount == 20
-
-
-@pytest.mark.usefixtures("db_session")
-class UbbleWorkflowTest:
-    def test_start_ubble_workflow(self, ubble_mock):
-        user = users_factories.UserFactory()
-        redirect_url = subscription_api.start_ubble_workflow(user, redirect_url="https://example.com")
-        assert redirect_url == "https://id.ubble.ai/29d9eca4-dce6-49ed-b1b5-8bb0179493a8"
-
-        fraud_check = user.beneficiaryFraudChecks[0]
-        assert fraud_check.type == fraud_models.FraudCheckType.UBBLE
-        assert fraud_check.thirdPartyId is not None
-        assert fraud_check.resultContent is not None
-        assert fraud_check.status == fraud_models.FraudCheckStatus.PENDING
-
-        ubble_request = ubble_mock.last_request.json()
-        assert ubble_request["data"]["attributes"]["webhook"] == "http://localhost/webhooks/ubble/application_status"
-
-    @pytest.mark.parametrize(
-        "state, status, fraud_check_status",
-        [
-            (
-                IdentificationState.INITIATED,
-                fraud_models.ubble.UbbleIdentificationStatus.INITIATED,
-                fraud_models.FraudCheckStatus.PENDING,
-            ),
-            (
-                IdentificationState.PROCESSING,
-                fraud_models.ubble.UbbleIdentificationStatus.PROCESSING,
-                fraud_models.FraudCheckStatus.PENDING,
-            ),
-            (
-                IdentificationState.VALID,
-                fraud_models.ubble.UbbleIdentificationStatus.PROCESSED,
-                fraud_models.FraudCheckStatus.OK,
-            ),
-            (
-                IdentificationState.INVALID,
-                fraud_models.ubble.UbbleIdentificationStatus.PROCESSED,
-                fraud_models.FraudCheckStatus.KO,
-            ),
-            (
-                IdentificationState.UNPROCESSABLE,
-                fraud_models.ubble.UbbleIdentificationStatus.PROCESSED,
-                fraud_models.FraudCheckStatus.KO,
-            ),
-            (
-                IdentificationState.ABORTED,
-                fraud_models.ubble.UbbleIdentificationStatus.ABORTED,
-                fraud_models.FraudCheckStatus.CANCELED,
-            ),
-        ],
-    )
-    def test_update_ubble_workflow(self, ubble_mocker, state, status, fraud_check_status):
-        user = users_factories.UserFactory()
-        fraud_check = fraud_factories.BeneficiaryFraudCheckFactory(type=fraud_models.FraudCheckType.UBBLE, user=user)
-        ubble_response = UbbleIdentificationResponseFactory(identification_state=state)
-
-        with ubble_mocker(
-            fraud_check.thirdPartyId,
-            json.dumps(ubble_response.dict(by_alias=True), sort_keys=True, default=json_default),
-        ):
-            subscription_api.update_ubble_workflow(fraud_check, ubble_response.data.attributes.status)
-
-        ubble_content = fraud_check.resultContent
-        assert ubble_content["status"] == status.value
-        assert fraud_check.status == fraud_check_status
-
-    def test_ubble_workflow_processing_add_inapp_message(self, ubble_mocker):
-        user = users_factories.UserFactory()
-        fraud_check = fraud_factories.BeneficiaryFraudCheckFactory(
-            type=fraud_models.FraudCheckType.UBBLE, status=fraud_models.FraudCheckStatus.PENDING, user=user
-        )
-        ubble_response = UbbleIdentificationResponseFactory(identification_state=IdentificationState.PROCESSING)
-        with ubble_mocker(
-            fraud_check.thirdPartyId,
-            json.dumps(ubble_response.dict(by_alias=True), sort_keys=True, default=json_default),
-        ):
-            subscription_api.update_ubble_workflow(fraud_check, ubble_response.data.attributes.status)
-            message = subscription_models.SubscriptionMessage.query.one()
-            assert message.userMessage == "Ton document d'identité est en cours de vérification."
-            assert message.popOverIcon == subscription_models.PopOverIcon.CLOCK
-
-    def test_ubble_workflow_rejected_add_inapp_message(self, ubble_mocker):
-        user = users_factories.UserFactory()
-        fraud_check = fraud_factories.BeneficiaryFraudCheckFactory(
-            type=fraud_models.FraudCheckType.UBBLE, status=fraud_models.FraudCheckStatus.PENDING, user=user
-        )
-        ubble_response = UbbleIdentificationResponseFactory(identification_state=IdentificationState.INVALID)
-        with ubble_mocker(
-            fraud_check.thirdPartyId,
-            json.dumps(ubble_response.dict(by_alias=True), sort_keys=True, default=json_default),
-        ):
-            subscription_api.update_ubble_workflow(fraud_check, ubble_response.data.attributes.status)
-            message = subscription_models.SubscriptionMessage.query.one()
-            assert (
-                message.userMessage
-                == "Désolé, la vérification de ton identité n'a pas pu aboutir. Nous t'invitons à passer par le site Démarches-Simplifiées."
-            )
-            assert message.callToActionLink == "passculture://verification-identite/demarches-simplifiees"
-            assert message.callToActionIcon == subscription_models.CallToActionIcon.EXTERNAL
-            assert message.callToActionTitle == "Accéder au site Démarches-Simplifiées"
 
 
 @pytest.mark.usefixtures("db_session")
